@@ -1,8 +1,8 @@
 from enum import Enum
 from re import findall, fullmatch
-from typing import Dict, List, Optional
+from typing import Any, Dict, List, Optional
 
-from pydantic import BaseModel, Field, StrictStr, conint, validator
+from pydantic import BaseModel, Field, StrictStr, validator
 
 from .metas.Metas import Speaker, SpeakerInfo
 
@@ -25,6 +25,9 @@ class Mora(BaseModel):
             for k, v in self.__dict__.items()
         ]
         return hash(tuple(sorted(items)))
+
+    class Config:
+        validate_assignment = True
 
 
 class AccentPhrase(BaseModel):
@@ -59,7 +62,7 @@ class AudioQuery(BaseModel):
     postPhonemeLength: float = Field(title="音声の後の無音時間")
     outputSamplingRate: int = Field(title="音声データの出力サンプリングレート")
     outputStereo: bool = Field(title="音声データをステレオ出力するか否か")
-    kana: Optional[str] = Field(title="[読み取り専用]AquesTalk風記法によるテキスト。音声合成クエリとしては無視される")
+    kana: Optional[str] = Field(title="[読み取り専用]AquesTalk 風記法によるテキスト。音声合成用のクエリとしては無視される")
 
     def __hash__(self):
         items = [
@@ -67,6 +70,46 @@ class AudioQuery(BaseModel):
             for k, v in self.__dict__.items()
         ]
         return hash(tuple(sorted(items)))
+
+
+class Note(BaseModel):
+    """
+    音符ごとの情報
+    """
+
+    key: int | None = Field(title="音階")
+    frame_length: int = Field(title="音符のフレーム長")
+    lyric: str = Field(title="音符の歌詞")
+
+
+class Score(BaseModel):
+    """
+    楽譜情報
+    """
+
+    notes: List[Note] = Field(title="音符のリスト")
+
+
+class FramePhoneme(BaseModel):
+    """
+    音素の情報
+    """
+
+    phoneme: str = Field(title="音素")
+    frame_length: int = Field(title="音素のフレーム長")
+
+
+class FrameAudioQuery(BaseModel):
+    """
+    フレームごとの音声合成用のクエリ
+    """
+
+    f0: List[float] = Field(title="フレームごとの基本周波数")
+    volume: List[float] = Field(title="フレームごとの音量")
+    phonemes: List[FramePhoneme] = Field(title="音素のリスト")
+    volumeScale: float = Field(title="全体の音量")
+    outputSamplingRate: int = Field(title="音声データの出力サンプリングレート")
+    outputStereo: bool = Field(title="音声データをステレオ出力するか否か")
 
 
 class ParseKanaErrorCode(Enum):
@@ -80,10 +123,10 @@ class ParseKanaErrorCode(Enum):
 
 
 class ParseKanaError(Exception):
-    def __init__(self, errcode: ParseKanaErrorCode, **kwargs):
+    def __init__(self, errcode: ParseKanaErrorCode, **kwargs: Any) -> None:
         self.errcode = errcode
         self.errname = errcode.name
-        self.kwargs: Dict[str, str] = kwargs
+        self.kwargs = kwargs
         err_fmt: str = errcode.value
         self.text = err_fmt.format(**kwargs)
 
@@ -167,8 +210,8 @@ class UserDictWord(BaseModel):
     """
 
     surface: str = Field(title="表層形")
-    priority: conint(ge=USER_DICT_MIN_PRIORITY, le=USER_DICT_MAX_PRIORITY) = Field(
-        title="優先度"
+    priority: int = Field(
+        title="優先度", ge=USER_DICT_MIN_PRIORITY, le=USER_DICT_MAX_PRIORITY
     )
     context_id: int = Field(title="文脈ID", default=1348)
     part_of_speech: str = Field(title="品詞")
